# docs/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/units": {
      "get": {
        "operationId": "UnitsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 160,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUnitsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Units"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "UnitsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Units"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/units/{id}": {
      "patch": {
        "operationId": "UnitsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentialUnit"
                }
              }
            }
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Units"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UnitsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Phase 0 error (400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "401": {
            "description": "Phase 0 error (401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "403": {
            "description": "Phase 0 error (403)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "404": {
            "description": "Phase 0 error (404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "409": {
            "description": "Phase 0 error (409)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "429": {
            "description": "Phase 0 error (429)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          },
          "500": {
            "description": "Phase 0 error (500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase0Error"
                }
              }
            }
          }
        },
        "tags": [
          "Units"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "SIGRA API",
    "description": "Official SIGRA Phase 0 HTTP contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "ResidentialUnit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "parkingSpaces": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "address",
          "parkingSpaces",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UnitResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "parkingSpaces": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "code",
          "address",
          "parkingSpaces",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedUnitsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateUnitDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30
          },
          "address": {
            "type": "string",
            "minLength": 5,
            "maxLength": 160
          },
          "parkingSpaces": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000
          }
        },
        "required": [
          "code",
          "address",
          "parkingSpaces"
        ]
      },
      "UpdateUnitDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30
          },
          "address": {
            "type": "string",
            "minLength": 5,
            "maxLength": 160
          },
          "parkingSpaces": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "Phase0Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}